+++ sdk/keyvault/azure-security-keyvault-keys/inc/azure/keyvault/keys/key_client_options.hpp
@@ -12,7 +12,7 @@
 #include <azure/core/internal/client_options.hpp>
 
 #include "azure/keyvault/keys/dll_import_export.hpp"
-#include "azure/keyvault/keys/key_vault_key.hpp"
+#include "azure/keyvault/keys/keyvault_key.hpp"
 
 namespace Azure { namespace Security { namespace KeyVault { namespace Keys {