+++ sdk/keyvault/azure-security-keyvault-keys/test/samples/sample5-sign-verify/sample5_sign_verify.cpp
@@ -19,8 +19,7 @@
 
 #include <azure/core.hpp>
 #include <azure/identity.hpp>
-#include <azure/keyvault/common/sha.hpp>
-#include <azure/keyvault/key_vault_keys.hpp>
+#include <azure/keyvault/keyvault_keys.hpp>
 
 #include <chrono>
 #include <iostream>
@@ -60,51 +59,54 @@ int main()
 
   CryptographyClient ecCryptoClient(cloudEcKey.Id(), credential);
 
-  // digestBase64 simulates some text data that has been hashed using the SHA256 algorithm
+  uint8_t const dataSource[]
+      = "This is some sample data which we will use to demonstrate sign and verify";
+  std::vector<uint8_t> data(std::begin(dataSource), std::end(dataSource));
+
+  // digestRaw simulates some text data that has been hashed using the SHA256 algorithm
   // and then base 64 encoded. It is not relevant for the sample how to create the SHA256
   // hashed digest.
-  // Example input data source for the digest:
-  // "This is some sample data which we will use to demonstrate sign and verify"
-  std::string digestBase64 = "DU9EdhpwhJqnGnieD0qKYEz6e8QPKlOVpYZZro+XtI8=";
-  std::vector<uint8_t> digest = Azure::Core::Convert::Base64Decode(digestBase64);
+  std::vector<uint8_t> digest
+      = Azure::Core::Convert::Base64Decode("DU9EdhpwhJqnGnieD0qKYEz6e8QPKlOVpYZZro");
 
   // Sign and Verify from digest
-  SignResult rsaSignResult = rsaCryptoClient.Sign(SignatureAlgorithm::RS256, digest);
+  SignResult rsaSignResult = rsaCryptoClient.Sign(SignatureAlgorithm::RS256, digest).Value;
   std::cout << " - Signed digest using the algorithm " << rsaSignResult.Algorithm.ToString()
             << ", with key " << rsaSignResult.KeyId << ". The resulting signature is: "
             << Azure::Core::Convert::Base64Encode(rsaSignResult.Signature) << std::endl;
 
-  SignResult ecSignResult = ecCryptoClient.Sign(SignatureAlgorithm::ES256K, digest);
+  SignResult ecSignResult = ecCryptoClient.Sign(SignatureAlgorithm::ES256K, digest).Value;
   std::cout << " - Signed digest using the algorithm " << ecSignResult.Algorithm.ToString()
             << ", with key " << ecSignResult.KeyId << ". The resulting signature is: "
             << Azure::Core::Convert::Base64Encode(ecSignResult.Signature) << std::endl;
 
   VerifyResult rsaVerifyResult
-      = rsaCryptoClient.Verify(SignatureAlgorithm::RS256, digest, rsaSignResult.Signature);
+      = rsaCryptoClient.Verify(SignatureAlgorithm::RS256, digest, rsaSignResult.Signature).Value;
   std::cout << " - Verified the signature using the algorithm "
             << rsaVerifyResult.Algorithm.ToString() << ", with key " << rsaVerifyResult.KeyId
             << ". Signature is valid: " << (rsaVerifyResult.IsValid ? "True" : "False")
             << std::endl;
 
   VerifyResult ecVerifyResult
-      = ecCryptoClient.Verify(SignatureAlgorithm::ES256K, digest, ecSignResult.Signature);
+      = ecCryptoClient.Verify(SignatureAlgorithm::ES256K, digest, ecSignResult.Signature).Value;
   std::cout << " - Verified the signature using the algorithm "
             << ecVerifyResult.Algorithm.ToString() << ", with key " << ecVerifyResult.KeyId
             << ". Signature is valid: " << (ecVerifyResult.IsValid ? "True" : "False") << std::endl;
 
   // Sign and Verify from data
-  SignResult rsaSignDataResult = rsaCryptoClient.SignData(SignatureAlgorithm::RS256, data);
+  SignResult rsaSignDataResult = rsaCryptoClient.SignData(SignatureAlgorithm::RS256, data).Value;
   std::cout << " - Signed data using the algorithm " << rsaSignDataResult.Algorithm.ToString()
             << ", with key " << rsaSignDataResult.KeyId << ". The resulting signature is: "
             << Azure::Core::Convert::Base64Encode(rsaSignDataResult.Signature) << std::endl;
 
-  SignResult ecSignDataResult = ecCryptoClient.SignData(SignatureAlgorithm::ES256K, data);
+  SignResult ecSignDataResult = ecCryptoClient.SignData(SignatureAlgorithm::ES256K, data).Value;
   std::cout << " - Signed data using the algorithm " << ecSignDataResult.Algorithm.ToString()
             << ", with key " << ecSignDataResult.KeyId << ". The resulting signature is: "
             << Azure::Core::Convert::Base64Encode(ecSignDataResult.Signature) << std::endl;
 
   VerifyResult rsaVerifyDataResult
-      = rsaCryptoClient.VerifyData(SignatureAlgorithm::RS256, data, rsaSignDataResult.Signature);
+      = rsaCryptoClient.VerifyData(SignatureAlgorithm::RS256, data, rsaSignDataResult.Signature)
+            .Value;
   std::cout << " - Verified the signature using the algorithm "
             << rsaVerifyDataResult.Algorithm.ToString() << ", with key "
             << rsaVerifyDataResult.KeyId
@@ -112,7 +114,8 @@ int main()
             << std::endl;
 
   VerifyResult ecVerifyDataResult
-      = ecCryptoClient.VerifyData(SignatureAlgorithm::ES256K, data, ecSignDataResult.Signature);
+      = ecCryptoClient.VerifyData(SignatureAlgorithm::ES256K, data, ecSignDataResult.Signature)
+            .Value;
   std::cout << " - Verified the signature using the algorithm "
             << ecVerifyDataResult.Algorithm.ToString() << ", with key " << ecVerifyDataResult.KeyId
             << ". Signature is valid: " << (ecVerifyDataResult.IsValid ? "True" : "False")