+++ sdk/keyvault/azure-security-keyvault-secrets/src/secret_client.cpp
@@ -1,89 +1,17 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // SPDX-License-Identifier: MIT
-#pragma once
-#include "azure/keyvault/secrets/secret_client.hpp"
-#include "private/package_version.hpp"
-#include "private/secret_constants.hpp"
-#include "private/secret_serializers.hpp"
 
-#include <azure/core/credentials/credentials.hpp>
-#include <azure/core/http/http.hpp>
-#include <azure/core/http/policies/policy.hpp>
-#include <azure/keyvault/common/internal/keyvault_pipeline.hpp>
+/**
+ * @brief Keyvault Secrets Client definition.
+ *
+ */
+#include "azure/keyvault/secrets/secret_client.hpp"
+
+#include "private/package_version.hpp"
 
 #include <string>
 
 using namespace Azure::Security::KeyVault::Secrets;
-using namespace Azure::Core::Http::Policies;
-using namespace Azure::Core::Http::Policies::_internal;
+using namespace Azure::Security::KeyVault::Secrets::_detail;
 
-namespace {
-constexpr static const char TelemetryName[] = "keyvault-secrets";
-};
-
-std::string SecretClient::ClientVersion() const { return _detail::PackageVersion::ToString(); }
-
-SecretClient::SecretClient(
-    std::string const& vaultUrl,
-    std::shared_ptr<Core::Credentials::TokenCredential const> credential,
-    SecretClientOptions options)
-{
-  auto apiVersion = options.Version.ToString();
-
-  std::vector<std::unique_ptr<HttpPolicy>> perRetrypolicies;
-  {
-    Azure::Core::Credentials::TokenRequestContext const tokenContext
-        = {{"https://vault.azure.net/.default"}};
-
-    perRetrypolicies.emplace_back(
-        std::make_unique<BearerTokenAuthenticationPolicy>(credential, tokenContext));
-  }
-
-  m_pipeline = std::make_shared<Azure::Security::KeyVault::_internal::KeyVaultPipeline>(
-      Azure::Core::Url(vaultUrl),
-      apiVersion,
-      Azure::Core::Http::_internal::HttpPipeline(
-          options, TelemetryName, apiVersion, std::move(perRetrypolicies), {}));
-}
-
-Azure::Response<KeyVaultSecret> SecretClient::GetSecret(
-    std::string const& name,
-    GetSecretOptions const& options,
-    Azure::Core::Context const& context) const
-{
-  return m_pipeline->SendRequest<KeyVaultSecret>(
-      context,
-      Azure::Core::Http::HttpMethod::Get,
-      [&name](Azure::Core::Http::RawResponse const& rawResponse) {
-        return _detail::KeyVaultSecretSerializer::KeyVaultSecretDeserialize(name, rawResponse);
-      },
-      {_detail::SecretPath, name, options.Version});
-}
-/*
-Azure::Response<KeyVaultSecret> SecretClient::SetSecret(
-    std::string const& name,
-    std::string const& value,
-    Azure::Core::Context const& context) const
-{
-  KeyVaultSecret secret;
-  secret.Value = value;
-  secret.Attributes.Name = name;
-
-  return SetSecret(secret, context);
-}
-
-Azure::Response<KeyVaultSecret> SecretClient::SetSecret(
-    KeyVaultSecret const& secret,
-    Azure::Core::Context const& context) const
-{
-    
-  return m_pipeline->SendRequest<KeyVaultSecret>(
-      context,
-      Azure::Core::Http::HttpMethod::Put,
-      [](Azure::Core::Http::RawResponse const& rawResponse) {
-        return _detail::KeyVaultSecretSerializer::KeyVaultSecretDeserialize(rawResponse);
-      },
-      {_detail::SecretPath, secret.Attributes.Name});
-}
-*/
-const ServiceVersion ServiceVersion::V7_2("7.2");
+std::string SecretClient::ClientVersion() const { return PackageVersion::ToString(); }