+++ sdk/storage/azure-storage-queues/inc/azure/storage/queues.hpp
@@ -1,0 +1,9 @@
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// SPDX-License-Identifier: MIT
+
+/**
+ * @brief Includes all public headers from Azure Storage Queues SDK library.
+ *
+ */
+
+#pragma once